Add roomSchema file to separate schema from my seed connection
